+++ platforms/fabric-1.16/src/main/java/org/anti_ad/mc/ipnext/mixin/MixinPlayerInteractionManagerForLockedSlotsMovePrevention.java
@@ -6,6 +6,7 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.screen.slot.SlotActionType;
 import org.anti_ad.mc.ipnext.config.ModSettings;
 import org.anti_ad.mc.ipnext.event.LockSlotsHandler;
+import org.anti_ad.mc.ipnext.event.LockedSlotKeeper;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -26,8 +27,8 @@ public class MixinPlayerInteractionManagerForLockedSlotsMovePrevention {
                           PlayerEntity player,
                           CallbackInfoReturnable<ItemStack> cir) {
         if(ModSettings.INSTANCE.getLOCKED_SLOTS_DISABLE_QUICK_MOVE_THROW().getValue() &&
-                (actionType == SlotActionType.QUICK_MOVE ||
-                        (actionType == SlotActionType.THROW && button == 1))) {
+                !LockedSlotKeeper.INSTANCE.getProcessingLockedPickups() &&
+                (actionType == SlotActionType.QUICK_MOVE || (actionType == SlotActionType.THROW && button == 1))) {
             if (!LockSlotsHandler.INSTANCE.isQMoveActionAllowed(slotIndex, button)) {
                 cir.cancel();
             }