+++ platforms/fabric-1.17/src/main/java/org/anti_ad/mc/ipnext/event/ClientEventHandler.kt
@@ -24,17 +24,21 @@ object ClientEventHandler {
     }
 
     private fun onTickInGame() {
+        LockedSlotKeeper.onTickInGame()
+
         if (GuiSettings.ENABLE_INVENTORY_BUTTONS.booleanValue
             && GuiSettings.SHOW_CONTINUOUS_CRAFTING_CHECKBOX.booleanValue) {
-
             ContinuousCraftingHandler.onTickInGame()
         }
+
         if (ModSettings.ENABLE_AUTO_REFILL.booleanValue) {
             AutoRefillHandler.onTickInGame()
         }
+
         if (Tweaks.CONTAINER_SWIPE_MOVING_ITEMS.booleanValue) {
             MiscHandler.swipeMoving()
         }
+
         LockSlotsHandler.onTickInGame()
     }
 
@@ -42,6 +46,9 @@ object ClientEventHandler {
         GlobalInputHandler.pressedKeys.clear() // sometimes left up not captured
         if (ModSettings.ENABLE_AUTO_REFILL.booleanValue) {
             AutoRefillHandler.onJoinWorld()
+        }
+        if (ModSettings.ENABLE_LOCK_SLOTS.booleanValue && !ModSettings.LOCKED_SLOTS_ALLOW_PICKUP_INTO_EMPTY.booleanValue) {
+            LockedSlotKeeper.onJoinWorld()
         }
         CustomDataFileLoader.reload(clientWorld)
     }