+++ platforms/fabric-1.17/src/main/java/org/anti_ad/mc/ipnext/event/LockSlotsHandler.kt
@@ -159,8 +159,7 @@ object LockSlotsHandler {
                 if ((mode == 0) == (invSlot !in lockedInvSlotsStoredValue)
                     && line.intersects(Rectangle(topLeft + slotTopLeft,
                                                  Size(18,
-                                                      18)))
-                ) {
+                                                      18)))) {
                     if (mode == 0)
                         lockedInvSlotsStoredValue.add(invSlot)
                     else