+++ platforms/forge-1.16/src/main/java/org/anti_ad/mc/ipnext/mixin/MixinScreenHandler.java
@@ -8,6 +8,7 @@ import net.minecraft.inventory.container.PlayerContainer;
 import net.minecraft.item.ItemStack;
 import org.anti_ad.mc.ipnext.config.ModSettings;
 import org.anti_ad.mc.ipnext.event.LockSlotsHandler;
+import org.anti_ad.mc.ipnext.event.LockedSlotKeeper;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -27,8 +28,8 @@ public class MixinScreenHandler {
                                          PlayerEntity player,
                                          CallbackInfoReturnable<ItemStack> cir) {
         if(ModSettings.INSTANCE.getLOCKED_SLOTS_DISABLE_QUICK_MOVE_THROW().getValue() &&
-                (actionType == ClickType.QUICK_MOVE ||
-                (actionType == ClickType.THROW && button == 1))) {
+                !LockedSlotKeeper.INSTANCE.getProcessingLockedPickups() &&
+                (actionType == ClickType.QUICK_MOVE || (actionType == ClickType.THROW && button == 1))) {
             if (!LockSlotsHandler.INSTANCE.isQMoveActionAllowed(slotIndex, button)) {
               cir.cancel();
             }